Split monolithic cli.py into per-resource CLI modules
- Convert 9 single-file resources to packages (draft, gcal, contacts,
  gdrive, gslides, gtask, form, label, filter) preserving git history
- Create cli.py in each of 12 resource packages with their Click commands
- Create cli_lib.py with shared helpers (handle_errors, _confirm_and_push)
  and re-exports from ui.py
- Slim main cli.py from 2500 to 552 lines (top-level commands, auth,
  utilities only)
- All imports preserved via __init__.py re-exports
