+++ gax/cli.py
@@ -1,87 +1,32 @@
 """CLI interface for gax.
 
-Policy: All Click command definitions and CLI UX logic live here.
-
-Resource modules (draft.py, gcal.py, etc.) contain pure business logic
-and must not import Click or call sys.exit(). They communicate via:
-
-  - logging.info() / logging.debug()  — status messages (shown in spinner)
-  - ValueError                        — user-fixable errors
-  - Return values                     — results for cli.py to format
-
-Confirmation prompts (--yes, diff display) are handled here in cli.py
-using Resource.diff() to preview changes before calling push/pull.
-
-Output conventions for resource methods:
-  - No output (most ops): return None, cli.py prints success()
-  - Structured result (path, ID): return it, cli.py formats
-  - Tabular/streaming (list, diff): accept a file descriptor, write to it
-
-Imports: cli.py imports only the resource *class* (e.g. Draft, not
-parse_draft or create_draft). All interaction goes through class methods.
-Non-standard CLI commands can add methods to the class as needed.
-Module-to-module imports (e.g. mail.py using draft internals) are fine.
+Top-level commands (pull, push, clone, checkout) and infrastructure.
+Resource-specific commands live in each resource's cli.py module.
 """
 
-import functools
 import glob
 import sys
 import click
 from pathlib import Path
 
-from .gsheet import pull_all
 from . import auth
 from . import docs
-from .mail import Thread, Mailbox
-from .label import Label
-from .filter import Filter
-from .gcal import Cal, Event
-from .gtask import TaskList, Task as TaskResource
-from .form import Form
-from .draft import Draft
-from .contacts import Contact, Contacts  # noqa: F401 — registers with Resource._subclasses
+from .cli_lib import handle_errors, _confirm_and_push  # noqa: F401
 from .resource import Resource
-from .gdrive import File
-from .gdoc import Tab, Doc  # noqa: F401 — registers with Resource._subclasses
-from .gslides import Slide, Presentation  # noqa: F401
 
-
-# =============================================================================
-# CLI helpers — keep UI concerns out of Resource classes
-# =============================================================================
-
-
-def handle_errors(fn):
-    """Decorator: catch exceptions, print error, exit 1."""
-
-    @functools.wraps(fn)
-    def wrapper(*args, **kwargs):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as e:
-            from .ui import error
-
-            error(str(e))
-            sys.exit(1)
-
-    return wrapper
-
-
-def _confirm_and_push(resource, *, yes=False, **kw):
-    """Standard diff → confirm → push flow."""
-    from .ui import success
-
-    diff_text = resource.diff(**kw)
-    if diff_text is None:
-        click.echo("No changes to push.")
-        return
-    if not yes:
-        click.echo(diff_text)
-        if not click.confirm("Push these changes?"):
-            click.echo("Cancelled.")
-            return
-    resource.push(**kw)
-    success("Pushed successfully.")
+# Import resource CLI groups — triggers Resource.__init_subclass__ registration
+from .gsheet.cli import sheet
+from .draft.cli import draft
+from .contacts.cli import contacts
+from .gdrive.cli import file_group
+from .mail.cli import mail_group, mailbox_group
+from .label.cli import mail_label
+from .filter.cli import mail_filter
+from .gcal.cli import cal_group
+from .gtask.cli import task_group
+from .form.cli import form
+from .gdoc.cli import doc
+from .gslides.cli import slides
 
 
 @click.group()
@@ -91,6 +36,11 @@ def main():
     from . import ui
 
     ui.setup_logging()
+
+
+# =============================================================================
+# Top-level commands (dispatch via Resource registry)
+# =============================================================================
 
 
 @docs.section("main")
@@ -359,7 +309,9 @@ def man(ctx, md: bool):
         click.echo(format_man_plain(sections))
 
 
-# --- Auth commands ---
+# =============================================================================
+# Auth commands
+# =============================================================================
 
 
 @docs.section("utility")
@@ -416,1967 +368,11 @@ def logout():
         click.echo("No token to remove.")
 
 
-# --- GSheet commands ---
-
-
-def _find_sheet_folder() -> Path:
-    """Find a .sheet.gax.md.d folder in the current directory."""
-    candidates = list(Path.cwd().glob("*.sheet.gax.md.d"))
-    if len(candidates) == 0:
-        raise ValueError("No .sheet.gax.md.d folder found in current directory")
-    elif len(candidates) > 1:
-        names = ", ".join(c.name for c in candidates)
-        raise ValueError(f"Multiple .sheet.gax.md.d folders found: {names}")
-    return candidates[0]
-
-
-@docs.section("resource")
-@main.group()
-def sheet():
-    """Google Sheets operations"""
-    pass
-
-
-@sheet.command("clone")
-@click.argument("url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <title>.sheet.gax.md)",
-)
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    default="md",
-    help="Output format: md, csv, tsv, psv, json, jsonl",
-)
-@click.option(
-    "-q",
-    "--quiet",
-    is_flag=True,
-    help="Suppress multi-tab status message",
-)
-@handle_errors
-def sheet_clone(url: str, output: Path | None, fmt: str, quiet: bool):
-    """Clone first tab from a spreadsheet to a .sheet.gax.md file.
-
-    For all tabs, use 'gax sheet checkout'.
-    """
-    from .gsheet import SheetTab, _extract_spreadsheet_id
-    from .gsheet.client import GSheetClient
-
-    file_path = SheetTab.from_url(url).clone(output=output, fmt=fmt)
-    click.echo(f"Created: {file_path}")
-
-    if not quiet:
-        spreadsheet_id = _extract_spreadsheet_id(url)
-        info = GSheetClient().get_spreadsheet_info(spreadsheet_id)
-        if len(info["tabs"]) > 1:
-            first_tab = info["tabs"][0]["title"]
-            click.echo(
-                f'  Tab "{first_tab}" cloned (1 of {len(info["tabs"])} tabs).\n'
-                f"  For all tabs: gax sheet checkout {url}"
-            )
-
-
-@sheet.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def sheet_pull(file: Path):
-    """Pull latest data for all tabs in a multipart file or checkout folder."""
-    from .ui import success as ui_success
-
-    if file.is_dir():
-        from .gsheet import Sheet
-
-        Sheet(path=file).pull()
-        ui_success(f"Pulled: {file}")
-    else:
-        rows = pull_all(file)
-        ui_success(f"Pulled {rows} rows to {file}")
-
-
-@sheet.command("checkout")
-@click.argument("url")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output folder (default: <title>.sheet.gax.md.d)",
-)
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    default="md",
-    help="Output format: md, csv, tsv, psv, json, jsonl",
-)
-@handle_errors
-def sheet_checkout(url: str, output: Path | None, fmt: str):
-    """Checkout all tabs to individual files in a folder.
-
-    Creates a folder with individual .tab.sheet.gax.md files for each tab.
-    Incremental: skips existing files.
-
-    \b
-    Examples:
-        gax sheet checkout <url>
-        gax sheet checkout <url> -o MyBudget/
-        gax sheet checkout <url> -f csv
-    """
-    from .gsheet import Sheet
-    from .ui import success
-
-    folder = Sheet.from_url(url).checkout(output=output, fmt=fmt)
-    success(f"Checked out to: {folder}")
-
-
-@sheet.command("push")
-@click.argument("folder", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "--with-formulas", is_flag=True, help="Interpret formulas (e.g. =SUM(A1:A10))"
-)
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation prompt")
-@handle_errors
-def sheet_push(folder: Path, with_formulas: bool, yes: bool):
-    """Push all tabs in a checkout folder to Google Sheets.
-
-    Shows a diff preview of changes and prompts for confirmation before pushing.
-
-    \b
-    Examples:
-        gax sheet push Budget.sheet.gax.md.d
-        gax sheet push Budget.sheet.gax.md.d -y
-        gax sheet push Budget.sheet.gax.md.d --with-formulas
-    """
-    from .gsheet import Sheet
-
-    _confirm_and_push(Sheet(path=folder), yes=yes, with_formulas=with_formulas)
-
-
-@sheet.command("plan")
-@click.argument("folder", type=click.Path(exists=True, path_type=Path), required=False)
-@handle_errors
-def sheet_plan(folder):
-    """Show what changes would be pushed to Google Sheets.
-
-    Similar to 'terraform plan' - previews changes without applying them.
-    If no folder is specified, looks for a .sheet.gax.md.d folder in the current directory.
-
-    \b
-    Examples:
-        gax sheet plan
-        gax sheet plan Budget.sheet.gax.md.d
-    """
-    from .gsheet import Sheet
-
-    if folder is None:
-        folder = _find_sheet_folder()
-
-    diff_text = Sheet(path=folder).diff()
-    if diff_text is None:
-        click.echo("No changes to push.")
-    else:
-        click.echo("\n" + diff_text)
-        click.echo(
-            "\nRun 'gax sheet apply' to push these changes, or 'gax sheet push <folder>' with confirmation."
-        )
-
-
-@sheet.command("apply")
-@click.argument("folder", type=click.Path(exists=True, path_type=Path), required=False)
-@click.option(
-    "--with-formulas", is_flag=True, help="Interpret formulas (e.g. =SUM(A1:A10))"
-)
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def sheet_apply(folder, with_formulas: bool, yes: bool):
-    """Apply planned changes by pushing to Google Sheets.
-
-    Similar to 'terraform apply' - shows plan and applies changes with confirmation.
-    If no folder is specified, looks for a .sheet.gax.md.d folder in the current directory.
-
-    \b
-    Examples:
-        gax sheet apply
-        gax sheet apply Budget.sheet.gax.md.d
-        gax sheet apply Budget.sheet.gax.md.d --with-formulas
-    """
-    from .gsheet import Sheet
-
-    if folder is None:
-        folder = _find_sheet_folder()
-
-    _confirm_and_push(Sheet(path=folder), yes=yes, with_formulas=with_formulas)
-
-
-@sheet.group()
-def tab():
-    """Single tab operations"""
-    pass
-
-
-@tab.command("list")
-@click.argument("url")
-@handle_errors
-def tab_list(url: str):
-    """List tabs in a spreadsheet (TSV output)."""
-    from .gsheet import Sheet
-
-    Sheet.from_url(url).tab_list(sys.stdout)
-
-
-@tab.command("clone")
-@click.argument("url")
-@click.argument("tab_name")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <tab>.sheet.gax.md)",
-)
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    default="md",
-    help="Output format: md, csv, tsv, psv, json, jsonl",
-)
-@handle_errors
-def tab_clone(url: str, tab_name: str, output: Path | None, fmt: str):
-    """Clone a single tab to a .sheet.gax.md file."""
-    from .gsheet import SheetTab
-    from .ui import success
-
-    path = SheetTab.from_url(url).clone(output=output, tab_name=tab_name, fmt=fmt)
-    success(f"Created: {path}")
-
-
-@tab.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def tab_pull(file: Path):
-    """Pull latest data for a single tab."""
-    from .gsheet import SheetTab
-    from .ui import success
-
-    SheetTab(path=file).pull()
-    success(f"Updated: {file}")
-
-
-@tab.command("push")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "--with-formulas", is_flag=True, help="Interpret formulas (e.g. =SUM(A1:A10))"
-)
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation prompt")
-@handle_errors
-def tab_push(file: Path, with_formulas: bool, yes: bool):
-    """Push local data to a single tab."""
-    from .gsheet import SheetTab
-    from .gsheet.frontmatter import parse_file
-    from .formats import get_format as get_fmt
-
-    config, data = parse_file(file)
-    fmt = get_fmt(config.format)
-    df = fmt.read(data)
-    row_count = len(df)
-
-    click.echo(f"Push {row_count} rows from {file} to {config.tab}?")
-    if not yes and not click.confirm("Proceed?"):
-        click.echo("Aborted.")
-        return
-
-    SheetTab(path=file).push(with_formulas=with_formulas)
-    click.echo(f"Pushed {row_count} rows")
-
-
 # =============================================================================
-# Draft commands
+# Register resource command groups
 # =============================================================================
 
-
-@docs.section("resource")
-@click.group()
-def draft():
-    """Draft operations"""
-    pass
-
-
-@draft.command("new")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <subject>.draft.gax.md)",
-)
-@click.option("--to", "to_addr", default="", help="Recipient email address")
-@click.option("--subject", default="", help="Email subject")
-@handle_errors
-def draft_new(output, to_addr, subject):
-    """Create a new local draft file.
-
-    Creates a .draft.gax.md file that can be edited and pushed to Gmail.
-
-    Examples:
-
-        gax draft new
-        gax draft new --to alice@example.com --subject "Hello"
-        gax draft new -o my_draft.draft.gax.md
-    """
-    from .ui import success
-
-    if not to_addr:
-        to_addr = click.prompt("To")
-    if not subject:
-        subject = click.prompt("Subject")
-
-    file_path = Draft().new(to=to_addr, subject=subject, output=output)
-    success(f"Created: {file_path}")
-    click.echo(f"Edit the file, then run: gax draft push {file_path}")
-
-
-@draft.command("clone")
-@click.argument("draft_id_or_url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <subject>.draft.gax.md)",
-)
-@handle_errors
-def draft_clone(draft_id_or_url, output):
-    """Clone an existing draft from Gmail.
-
-    Examples:
-
-        gax draft clone r-1234567890123456789
-        gax draft clone "https://mail.google.com/mail/u/0/#drafts/..."
-        gax draft clone r-1234567890 -o my_draft.draft.gax.md
-    """
-    from .ui import success
-
-    path = Draft.from_url_or_id(draft_id_or_url).clone(output=output)
-    success(f"Created: {path}")
-
-
-@draft.command("list")
-@click.option("--limit", default=100, help="Maximum results (default: 100)")
-@handle_errors
-def draft_list(limit):
-    """List Gmail drafts (TSV output).
-
-    Output columns: draft_id, thread_id, date, to, subject
-    """
-    Draft().list(sys.stdout, limit=limit)
-
-
-@draft.command("push")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation prompt")
-@handle_errors
-def draft_push(file, yes):
-    """Push local draft to Gmail.
-
-    If the draft doesn't exist in Gmail yet, creates it.
-    If it exists, shows diff and updates it (with confirmation).
-
-    Examples:
-
-        gax draft push my_draft.draft.gax.md
-        gax draft push my_draft.draft.gax.md -y
-    """
-    _confirm_and_push(Draft.from_file(file), yes=yes)
-
-
-@draft.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def draft_pull(file):
-    """Pull latest content from Gmail draft.
-
-    Updates the local .draft.gax.md file with the remote draft content.
-
-    Example:
-
-        gax draft pull my_draft.draft.gax.md
-    """
-    from .ui import success
-
-    Draft.from_file(file).pull()
-    success(f"Updated: {file}")
-
-
-# =============================================================================
-# Contacts commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group()
-def contacts():
-    """Google Contacts operations."""
-    pass
-
-
-@contacts.command("clone")
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    type=click.Choice(["md", "jsonl"]),
-    default="md",
-    help="Output format: md (view-only) or jsonl (editable)",
-)
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file (default: contacts.<format>)",
-)
-@handle_errors
-def contacts_clone(fmt, output):
-    """Clone all contacts to a local file.
-
-    \b
-    Formats:
-      md     Human-readable markdown (default, view-only)
-      jsonl  JSON Lines format (editable, scriptable)
-    """
-    from .ui import success
-
-    path = Contacts().clone(fmt=fmt, output=output)
-    success(f"Created: {path}")
-
-
-@contacts.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def contacts_pull(file):
-    """Pull latest contacts from Google.
-
-    Updates the file with current contact data, preserving format.
-    """
-    from .ui import success
-
-    Contacts(path=file).pull()
-    success(f"Updated: {file}")
-
-
-@contacts.command("push")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def contacts_push(file, yes):
-    """Push local JSONL contacts to Google.
-
-    Compares local contacts with remote, shows diff, and applies changes.
-    Only works with JSONL format files.
-    """
-    _confirm_and_push(Contacts(path=file), yes=yes)
-
-
-@contacts.command("checkout")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output folder (default: contacts.contacts.gax.md.d)",
-)
-@handle_errors
-def contacts_checkout(output):
-    """Checkout contacts as individual files into a folder.
-
-    Creates one .contact.gax.yaml file per contact for easy per-contact
-    editing and diffing.
-    """
-    from .ui import success
-
-    cloned, skipped = Contacts().checkout(output=output)
-    parts = [f"{cloned} contacts"]
-    if skipped:
-        parts.append(f"({skipped} skipped)")
-    success(f"Checked out {' '.join(parts)}")
-
-
-# =============================================================================
-# File commands (Google Drive)
-# =============================================================================
-
-
-@docs.section("resource")
-@docs.maturity("unstable")
-@click.group("file")
-def file_group():
-    """Google Drive file operations."""
-    pass
-
-
-@file_group.command("clone")
-@click.argument("url_or_id")
-@click.option(
-    "-o", "--output", type=click.Path(path_type=Path), help="Output file path"
-)
-@handle_errors
-def file_clone(url_or_id, output):
-    """Clone a file from Google Drive.
-
-    Downloads the file and creates a tracking .gax.md file.
-
-    Examples:
-
-        gax file clone https://drive.google.com/file/d/abc123/view
-        gax file clone abc123 -o report.pdf
-    """
-    from .ui import success
-
-    path = File.from_url_or_id(url_or_id).clone(output=output)
-    success(f"Created: {path}")
-
-
-@file_group.command("checkout")
-@click.argument("url_or_id")
-@click.option(
-    "-o", "--output", type=click.Path(path_type=Path), help="Output folder path"
-)
-@click.option("-R", "--recursive", is_flag=True, help="Recurse into subfolders")
-@handle_errors
-def file_checkout(url_or_id, output, recursive):
-    """Checkout a Google Drive folder to a local directory.
-
-    Downloads all files. Google Workspace files (Docs, Sheets, Forms)
-    are cloned via their native gax resource.
-
-    \b
-    Examples:
-        gax file checkout https://drive.google.com/drive/folders/abc123
-        gax file checkout abc123 -o my_folder
-        gax file checkout abc123 -R
-    """
-    from .ui import success
-    from .gdrive import Folder
-
-    path = Folder.from_url_or_id(url_or_id).checkout(output=output, recursive=recursive)
-    success(f"Checked out: {path}")
-
-
-@file_group.command("pull")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def file_pull(file_path):
-    """Pull latest version of a file from Google Drive.
-
-    Requires a .gax.md tracking file (created by 'gax file clone').
-
-    Example:
-
-        gax file pull report.pdf
-    """
-    from .ui import success
-
-    File(path=file_path).pull()
-    success(f"Updated: {file_path}")
-
-
-@file_group.command("push")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("--public", is_flag=True, help="Make file publicly accessible")
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def file_push(file_path, public, yes):
-    """Push local file to Google Drive.
-
-    If file has a .gax.md tracking file, updates existing file.
-    Otherwise, uploads as a new file.
-
-    Examples:
-
-        gax file push report.pdf
-        gax file push report.pdf --public
-        gax file push report.pdf -y
-    """
-    from .ui import success
-
-    tracking_path = file_path.with_suffix(file_path.suffix + ".gax.md")
-
-    if tracking_path.exists():
-        from .gdrive import read_tracking_file
-
-        tracking_data = read_tracking_file(tracking_path)
-        if not yes:
-            click.echo(f"Will update Drive file: {tracking_data.get('name')}")
-            click.echo(f"Local file: {file_path}")
-            if public:
-                click.echo("Will make publicly accessible")
-            if not click.confirm("Push these changes?"):
-                click.echo("Aborted.")
-                return
-    else:
-        if not yes:
-            click.echo(f"Will upload new file: {file_path.name}")
-            if public:
-                click.echo("Will make publicly accessible")
-            if not click.confirm("Upload this file?"):
-                click.echo("Aborted.")
-                return
-
-    File(path=file_path).push(public=public)
-    success("Pushed successfully.")
-
-
-# =============================================================================
-# Mail thread commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group()
-def mail_group():
-    """Individual email thread operations (clone, pull, reply)"""
-    pass
-
-
-@mail_group.command("clone")
-@click.argument("thread_id_or_url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file",
-)
-@handle_errors
-def mail_clone(thread_id_or_url, output):
-    """Clone a single email thread to a local .mail.gax.md file.
-
-    \b
-    Examples:
-        gax mail clone 19d0bed1cddbab6d
-        gax mail clone "https://mail.google.com/..."
-        gax mail clone 19d0bed1cddbab6d -o thread.mail.gax.md
-    """
-    from .ui import success
-
-    path = Thread.from_url_or_id(thread_id_or_url).clone(output=output)
-    success(f"Created: {path}")
-
-
-@mail_group.command("pull")
-@click.argument("path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def mail_pull(path):
-    """Pull latest messages for .mail.gax.md file(s).
-
-    Single file:
-
-        gax mail pull thread.mail.gax.md
-
-    Folder (updates all .mail.gax.md files):
-
-        gax mail pull Inbox/
-    """
-    from .ui import success
-
-    Thread(path=path).pull()
-    success(f"Updated: {path}")
-
-
-@mail_group.command("reply")
-@click.argument("file_or_url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: Re_<subject>.draft.gax.md)",
-)
-@handle_errors
-def mail_reply(file_or_url, output):
-    """Create a reply draft from a thread.
-
-    Examples:
-
-        gax mail reply Project_Update.mail.gax.md
-        gax mail reply "https://mail.google.com/mail/u/0/#inbox/abc123"
-        gax mail reply thread.mail.gax.md -o my_reply.draft.gax.md
-    """
-    from .ui import success
-
-    file_path = Path(file_or_url)
-    if file_path.exists():
-        thread = Thread(path=file_path)
-    else:
-        thread = Thread.from_url_or_id(file_or_url)
-    out_path = thread.reply(output=output)
-    success(f"Created: {out_path}")
-    click.echo(f"Edit the file, then run: gax draft push {out_path}")
-
-
-# =============================================================================
-# Mailbox commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group(invoke_without_command=True)
-@click.option(
-    "-q", "--query", default="in:inbox", help="Search query (default: in:inbox)"
-)
-@click.option("--limit", default=20, help="Maximum results (default: 20)")
-@click.pass_context
-@handle_errors
-def mailbox_group(ctx, query, limit):
-    """Search/list Gmail threads and bulk label operations.
-
-    Without subcommand, lists threads matching query (TSV output).
-
-    \b
-    Examples:
-        gax mailbox                        # List inbox
-        gax mailbox -q "from:alice"        # Search
-        gax mailbox clone                  # Clone for bulk labeling
-    """
-    if ctx.invoked_subcommand is None:
-        Mailbox().list(sys.stdout, query=query, limit=limit)
-
-
-@mailbox_group.command("fetch")
-@click.option(
-    "-o",
-    "--output",
-    default="mailbox.gax.md.d",
-    type=click.Path(path_type=Path),
-    help="Output folder (default: mailbox.gax.md.d)",
-)
-@click.option(
-    "-q", "--query", default="in:inbox", help="Search query (default: in:inbox)"
-)
-@click.option("--limit", default=50, help="Maximum threads (default: 50)")
-@handle_errors
-def mailbox_fetch(output, query, limit):
-    """Fetch full threads matching query into a folder."""
-    from .ui import success
-
-    cloned, skipped = Mailbox().fetch(query=query, limit=limit, output=output)
-    success(f"Cloned: {cloned}, Skipped: {skipped} (already present)")
-
-
-@mailbox_group.command("clone")
-@click.option(
-    "-o",
-    "--output",
-    default="mailbox.gax.md",
-    help="Output file (default: mailbox.gax.md)",
-)
-@click.option(
-    "-q", "--query", default="in:inbox", help="Search query (default: in:inbox)"
-)
-@click.option("--limit", default=50, help="Maximum threads (default: 50)")
-@handle_errors
-def mailbox_clone_cmd(output, query, limit):
-    """Clone threads from Gmail for bulk labeling."""
-    from .ui import success
-
-    file_path = Mailbox().clone(query=query, limit=limit, output=Path(output))
-    success(f"Cloned to: {file_path}")
-
-
-@mailbox_group.command("pull")
-@click.argument("file", type=click.Path(exists=True))
-@handle_errors
-def mailbox_pull(file):
-    """Update a .gax.md file by re-fetching from Gmail."""
-    from .ui import success
-
-    Mailbox(path=Path(file)).pull()
-    success(f"Updated: {file}")
-
-
-@mailbox_group.command("plan")
-@click.argument("file", type=click.Path(exists=True))
-@click.option(
-    "-o",
-    "--output",
-    default="mailbox.plan.yaml",
-    help="Output file (default: mailbox.plan.yaml)",
-)
-@handle_errors
-def mailbox_plan_cmd(file, output):
-    """Generate plan from edited list file."""
-    import yaml
-
-    plan = Mailbox(path=Path(file)).compute_plan()
-
-    if not plan["changes"]:
-        click.echo("No changes to apply.")
-        return
-
-    path = Path(output)
-    with open(path, "w") as f:
-        yaml.dump(
-            plan, f, default_flow_style=False, allow_unicode=True, sort_keys=False
-        )
-
-    changes = plan["changes"]
-    click.echo(f"Wrote {len(changes)} changes to {output}")
-
-    sys_add_count = sum(1 for c in changes if c.get("add_sys"))
-    sys_remove_count = sum(1 for c in changes if c.get("remove_sys"))
-    cat_change_count = sum(
-        1 for c in changes if c.get("add_cat") or c.get("remove_cat")
-    )
-    add_count = sum(1 for c in changes if c.get("add"))
-    remove_count = sum(1 for c in changes if c.get("remove"))
-
-    if sys_add_count or sys_remove_count:
-        click.echo(f"  System label changes: {sys_add_count + sys_remove_count}")
-    if cat_change_count:
-        click.echo(f"  Category changes: {cat_change_count}")
-    if add_count:
-        click.echo(f"  Add user labels: {add_count}")
-    if remove_count:
-        click.echo(f"  Remove user labels: {remove_count}")
-
-
-@mailbox_group.command("apply")
-@click.argument("plan_file", type=click.Path(exists=True))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def mailbox_apply(plan_file, yes):
-    """Apply label changes from plan."""
-    import yaml
-    from .ui import success, error
-
-    with open(plan_file) as f:
-        plan = yaml.safe_load(f)
-
-    changes = plan.get("changes", [])
-    if not changes:
-        click.echo("No changes in plan.")
-        return
-
-    click.echo(f"Plan: {plan_file}")
-    click.echo(f"Changes: {len(changes)}")
-    click.echo()
-
-    for change in changes[:10]:
-        thread_id = change["id"][:12] + "..."
-        actions = []
-        if change.get("add_sys"):
-            actions.append("+sys:" + ",".join(change["add_sys"]))
-        if change.get("remove_sys"):
-            actions.append("-sys:" + ",".join(change["remove_sys"]))
-        if change.get("add_cat"):
-            actions.append("+cat:" + change["add_cat"])
-        if change.get("remove_cat"):
-            actions.append("-cat:" + change["remove_cat"])
-        if change.get("add"):
-            actions.append("+" + ",".join(change["add"]))
-        if change.get("remove"):
-            actions.append("-" + ",".join(change["remove"]))
-        click.echo(f"  {thread_id}  {' '.join(actions)}")
-
-    if len(changes) > 10:
-        click.echo(f"  ... and {len(changes) - 10} more")
-
-    click.echo()
-
-    if not yes and not click.confirm("Apply these changes?"):
-        click.echo("Aborted.")
-        return
-
-    succeeded, failed = Mailbox().apply_plan(plan)
-    success(f"Applied: {succeeded} threads")
-    if failed:
-        error(f"Failed: {failed} threads")
-
-
-# =============================================================================
-# Label commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group("mail-label")
-def mail_label():
-    """Gmail label management (declarative)."""
-    pass
-
-
-@mail_label.command("list")
-@handle_errors
-def label_list():
-    """List Gmail labels (TSV output)."""
-    Label().list(sys.stdout)
-
-
-@mail_label.command("clone")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file (default: labels.mail.gax.md)",
-)
-@click.option("--all", "include_all", is_flag=True, help="Include system labels")
-@handle_errors
-def label_clone(output, include_all):
-    """Clone Gmail labels to a .gax.md file."""
-    from .ui import success
-
-    file_path = Label().clone(output=output, include_all=include_all)
-    success(f"Created: {file_path}")
-
-
-@mail_label.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("--all", "include_all", is_flag=True, help="Include system labels")
-@handle_errors
-def label_pull(file, include_all):
-    """Pull latest labels to existing file."""
-    from .ui import success
-
-    Label.from_file(file).pull(include_all=include_all)
-    success(f"Updated: {file}")
-
-
-@mail_label.command("plan")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    default="labels.plan.yaml",
-    help="Output plan file",
-)
-@click.option("--delete", "allow_delete", is_flag=True, help="Include deletions")
-@handle_errors
-def label_plan(file, output, allow_delete):
-    """Preview label changes (diff)."""
-    diff_text = Label.from_file(file).diff(allow_delete=allow_delete)
-    if diff_text is None:
-        click.echo("No changes to apply.")
-        return
-    click.echo(diff_text)
-
-
-@mail_label.command("apply")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@click.option("--delete", "allow_delete", is_flag=True, help="Include deletions")
-@handle_errors
-def label_apply(file, yes, allow_delete):
-    """Apply label changes to Gmail."""
-    _confirm_and_push(Label.from_file(file), yes=yes, allow_delete=allow_delete)
-
-
-# =============================================================================
-# Filter commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group("mail-filter")
-def mail_filter():
-    """Gmail filter management (declarative).
-
-    Note: Gmail applies ALL matching filters simultaneously, not sequentially.
-    Filter order has no significance - there is no "stop processing" feature.
-    """
-    pass
-
-
-@mail_filter.command("list")
-@handle_errors
-def filter_list():
-    """List Gmail filters (TSV output)."""
-    Filter().list(sys.stdout)
-
-
-@mail_filter.command("clone")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file (default: filters.mail.gax.md)",
-)
-@handle_errors
-def filter_clone(output):
-    """Clone Gmail filters to a .gax.md file."""
-    from .ui import success
-
-    file_path = Filter().clone(output=output)
-    success(f"Created: {file_path}")
-
-
-@mail_filter.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def filter_pull(file):
-    """Pull latest filters to existing file."""
-    from .ui import success
-
-    Filter.from_file(file).pull()
-    success(f"Updated: {file}")
-
-
-@mail_filter.command("plan")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def filter_plan(file):
-    """Preview filter changes (diff)."""
-    diff_text = Filter.from_file(file).diff()
-    if diff_text is None:
-        click.echo("No changes to apply.")
-        return
-    click.echo(diff_text)
-
-
-@mail_filter.command("apply")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def filter_apply(file, yes):
-    """Apply filter changes to Gmail."""
-    _confirm_and_push(Filter.from_file(file), yes=yes)
-
-
-# =============================================================================
-# Calendar commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group(name="cal")
-def cal_group():
-    """Google Calendar sync commands."""
-    pass
-
-
-@cal_group.command(name="calendars")
-@handle_errors
-def cal_calendars_cmd():
-    """List available calendars."""
-    Cal().calendars(sys.stdout)
-
-
-@cal_group.command(name="list")
-@click.argument("calendar", required=False)
-@click.option(
-    "--days", "-d", default=None, type=int, help="Number of days to show (default: 7)"
-)
-@click.option("--from", "date_from", default=None, help="Start date (YYYY-MM-DD)")
-@click.option("--to", "date_to", default=None, help="End date (YYYY-MM-DD)")
-@click.option(
-    "--format",
-    "-f",
-    "fmt",
-    type=click.Choice(["md", "tsv"]),
-    default="md",
-    help="Output format (default: md)",
-)
-@click.option("-v", "--verbose", is_flag=True, help="Include event descriptions")
-@handle_errors
-def cal_list_cmd(
-    calendar: str | None,
-    days: int | None,
-    date_from: str | None,
-    date_to: str | None,
-    fmt: str,
-    verbose: bool,
-):
-    """List events from a calendar.
-
-    CALENDAR is a calendar name, ID, or numeric index (from 'gax cal calendars').
-    Defaults to the primary calendar.
-
-    \b
-    Examples:
-        gax cal list                  # Primary calendar, next 7 days
-        gax cal list -d 14            # Next 14 days
-        gax cal list Work             # "Work" calendar
-        gax cal list --from 2026-03-01 --to 2026-03-15
-        gax cal list -f tsv           # TSV output
-    """
-    from .gcal import (
-        resolve_time_range,
-        resolve_calendar_id,
-        list_events,
-        format_events_tsv,
-        format_events_markdown,
-    )
-
-    time_min, time_max = resolve_time_range(days, date_from, date_to)
-    calendar_id = resolve_calendar_id(calendar)
-    events = list_events(
-        time_min=time_min, time_max=time_max, calendar_id=calendar_id
-    )
-
-    if fmt == "tsv":
-        click.echo(format_events_tsv(events, include_desc=verbose), nl=False)
-    else:
-        click.echo(format_events_markdown(events, include_desc=verbose), nl=False)
-
-
-@cal_group.command(name="clone")
-@click.argument("calendar", required=False)
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file (default: calendar.cal.gax.md)",
-)
-@click.option(
-    "--days", "-d", default=None, type=int, help="Number of days (default: 7)"
-)
-@click.option("--from", "date_from", default=None, help="Start date (YYYY-MM-DD)")
-@click.option("--to", "date_to", default=None, help="End date (YYYY-MM-DD)")
-@click.option("-v", "--verbose", is_flag=True, help="Include event descriptions")
-@handle_errors
-def cal_clone_cmd(
-    calendar: str | None,
-    output: Path | None,
-    days: int | None,
-    date_from: str | None,
-    date_to: str | None,
-    verbose: bool,
-):
-    """Clone events to a .cal.gax.md file.
-
-    Creates a file with all events that can be updated with 'gax cal pull'.
-    CALENDAR defaults to primary calendar.
-
-    \b
-    Examples:
-        gax cal clone
-        gax cal clone Work -o week.cal.gax.md -d 7
-        gax cal clone --from 2026-03-01 --to 2026-03-31 -o march.cal.gax.md
-    """
-    from .ui import success
-
-    file_path = Cal().clone(
-        output=output,
-        calendar=calendar,
-        days=days,
-        date_from=date_from,
-        date_to=date_to,
-        verbose=verbose,
-    )
-    success(f"Created: {file_path}")
-
-
-@cal_group.command(name="pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def cal_pull_cmd(file: Path):
-    """Pull latest events to existing file.
-
-    \b
-    Example:
-        gax cal pull week.cal.gax.md
-    """
-    from .ui import success
-
-    Cal(path=file).pull()
-    success(f"Updated: {file}")
-
-
-@cal_group.command(name="checkout")
-@click.argument("calendar", required=False)
-@click.option(
-    "-o",
-    "--output",
-    default="calendar.cal.gax.md.d",
-    type=click.Path(path_type=Path),
-    help="Output folder (default: calendar.cal.gax.md.d)",
-)
-@click.option(
-    "--days", "-d", default=None, type=int, help="Number of days (default: 7)"
-)
-@click.option("--from", "date_from", default=None, help="Start date (YYYY-MM-DD)")
-@click.option("--to", "date_to", default=None, help="End date (YYYY-MM-DD)")
-@handle_errors
-def cal_checkout_cmd(
-    calendar: str | None,
-    output: Path,
-    days: int | None,
-    date_from: str | None,
-    date_to: str | None,
-):
-    """Checkout events as individual .cal.gax.md files into a folder.
-
-    Each event becomes a separate file that can be edited and pushed.
-    CALENDAR defaults to primary calendar.
-
-    \b
-    Examples:
-        gax cal checkout
-        gax cal checkout -o Week/
-        gax cal checkout Work -o Week/ -d 7
-    """
-    from .ui import success
-
-    folder = Cal().checkout(
-        output=output,
-        calendar=calendar,
-        days=days,
-        date_from=date_from,
-        date_to=date_to,
-    )
-    success(f"Checked out to: {folder}")
-
-
-@cal_group.group(name="event")
-def cal_event_group():
-    """Event operations (clone, new, pull, push, delete)."""
-    pass
-
-
-@cal_event_group.command(name="clone")
-@click.argument("id_or_url")
-@click.option(
-    "--cal", "-c", "calendar", default="primary", help="Calendar ID (default: primary)"
-)
-@click.option(
-    "-o",
-    "--output",
-    "output_path",
-    type=click.Path(path_type=Path),
-    help="Output file path",
-)
-@handle_errors
-def cal_event_clone_cmd(id_or_url: str, calendar: str, output_path: Path | None):
-    """Clone an event to a local .cal.gax.md file."""
-    from .ui import success
-
-    file_path = Event.from_url_or_id(id_or_url).clone(calendar=calendar, output=output_path)
-    success(f"Cloned event to {file_path}")
-
-
-@cal_event_group.command(name="new")
-@click.option(
-    "--cal", "-c", "calendar", default="primary", help="Calendar ID (default: primary)"
-)
-@click.option(
-    "-o",
-    "--output",
-    "output_path",
-    type=click.Path(path_type=Path),
-    help="Output file path",
-)
-@handle_errors
-def cal_event_new_cmd(calendar: str, output_path: Path | None):
-    """Create a new event file (edit and push to create upstream)."""
-    from .ui import success
-
-    file_path = Event().new(calendar=calendar, output=output_path)
-    success(f"Created event template at {file_path}")
-    click.echo(f"Edit the file, then run: gax cal event push {file_path}")
-
-
-@cal_event_group.command(name="pull")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def cal_event_pull_cmd(file_path: Path):
-    """Pull latest event data from API."""
-    from .ui import success
-
-    Event(path=file_path).pull()
-    success(f"Pulled latest data to {file_path}")
-
-
-@cal_event_group.command(name="push")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def cal_event_push_cmd(file_path: Path, yes: bool):
-    """Push local changes to API."""
-    from .ui import success
-
-    ev = Event(path=file_path)
-    diff_text = ev.diff()
-    if diff_text is None:
-        click.echo("No changes to push.")
-        return
-    if not yes:
-        click.echo(diff_text)
-        if not click.confirm("Push these changes?"):
-            click.echo("Cancelled.")
-            return
-
-    link = ev.push()
-    success(f"Pushed event: {link}")
-
-
-@cal_event_group.command(name="delete")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def cal_event_delete_cmd(file_path: Path, yes: bool):
-    """Delete event from calendar."""
-    from .gcal import yaml_to_event
-    from .ui import success
-
-    content = file_path.read_text()
-    local_event = yaml_to_event(content)
-
-    if not yes:
-        click.echo(f"Delete event '{local_event.title}' from calendar?")
-        click.echo("This will also delete the local file.")
-        if not click.confirm("Proceed?"):
-            click.echo("Cancelled.")
-            return
-
-    title = Event(path=file_path).delete()
-    success(f"Deleted event '{title}'")
-
-
-# =============================================================================
-# Task commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group(name="task")
-def task_group():
-    """Google Tasks sync commands."""
-    pass
-
-
-@task_group.command(name="lists")
-@handle_errors
-def task_lists_cmd():
-    """List available task lists."""
-    TaskList().lists(sys.stdout)
-
-
-@task_group.command(name="list")
-@click.argument("tasklist", required=False)
-@click.option("--all", "show_all", is_flag=True, help="Include completed tasks")
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    type=click.Choice(["md", "yaml"]),
-    default="md",
-    help="Output format (default: md)",
-)
-@handle_errors
-def task_list_cmd(tasklist: str | None, show_all: bool, fmt: str):
-    """View tasks from a task list."""
-    TaskList().list(sys.stdout, tasklist=tasklist, show_all=show_all, fmt=fmt)
-
-
-@task_group.command(name="clone")
-@click.argument("tasklist", required=False)
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file path",
-)
-@click.option("--all", "show_all", is_flag=True, help="Include completed tasks")
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    type=click.Choice(["md", "yaml"]),
-    default="md",
-    help="Output format (default: md)",
-)
-@handle_errors
-def task_clone_cmd(tasklist: str | None, output: Path | None, show_all: bool, fmt: str):
-    """Clone a task list to a single file."""
-    from .ui import success
-
-    path = TaskList().clone(
-        tasklist=tasklist, output=output, fmt=fmt, show_all=show_all
-    )
-    success(f"Created: {path}")
-
-
-@task_group.command(name="checkout")
-@click.argument("tasklist", required=False)
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output folder path",
-)
-@click.option("--all", "show_all", is_flag=True, help="Include completed tasks")
-@handle_errors
-def task_checkout_cmd(tasklist: str | None, output: Path | None, show_all: bool):
-    """Checkout a task list as a folder of individual task files."""
-    from .ui import success
-
-    cloned, skipped = TaskList().checkout(
-        tasklist=tasklist, output=output, show_all=show_all
-    )
-    success(f"Checked out: {cloned}, Skipped: {skipped}")
-
-
-@task_group.command(name="pull")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def task_pull_cmd(file_path: Path):
-    """Pull latest task data from API."""
-    from .ui import success
-
-    name = file_path.name
-    if name.endswith(".tasks.gax.md") or name.endswith(".tasks.gax.yaml"):
-        TaskList.from_file(file_path).pull()
-    else:
-        TaskResource.from_file(file_path).pull()
-    success(f"Updated: {file_path}")
-
-
-@task_group.command(name="new")
-@click.argument("title")
-@click.option(
-    "--tasklist",
-    help="Task list name, ID, or index (default: first list)",
-)
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file path",
-)
-@handle_errors
-def task_new_cmd(title: str, tasklist: str | None, output: Path | None):
-    """Create a new task on Google."""
-    from .ui import success
-
-    path = TaskResource().new(title, tasklist=tasklist, output=output)
-    success(f"Created: {path}")
-
-
-@task_group.command(name="diff")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def task_diff_cmd(file_path: Path):
-    """Show differences between local task and remote."""
-    diff_text = TaskResource.from_file(file_path).diff()
-    if diff_text is None:
-        click.echo("No changes.")
-    else:
-        click.echo(diff_text)
-
-
-@task_group.command(name="push")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def task_push_cmd(file_path: Path, yes: bool):
-    """Push local task changes to API."""
-    from .ui import success
-
-    t = TaskResource.from_file(file_path)
-    diff_text = t.diff()
-    if diff_text is None:
-        click.echo("No changes to push.")
-        return
-    if not yes:
-        click.echo(diff_text)
-        if not click.confirm("Push these changes?"):
-            click.echo("Cancelled.")
-            return
-    title = t.push()
-    success(f"Pushed: {title}")
-
-
-@task_group.command(name="done")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def task_done_cmd(file_path: Path, yes: bool):
-    """Mark a task as completed and push."""
-    from .ui import success
-
-    if not yes:
-        if not click.confirm(f"Mark {file_path.name} as done?"):
-            click.echo("Cancelled.")
-            return
-    title = TaskResource.from_file(file_path).done()
-    success(f"Done: {title}")
-
-
-@task_group.command(name="delete")
-@click.argument("file_path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def task_delete_cmd(file_path: Path, yes: bool):
-    """Delete a task from Google and local file."""
-    from .ui import success
-
-    if not yes:
-        if not click.confirm(f"Delete {file_path.name}?"):
-            click.echo("Cancelled.")
-            return
-    title = TaskResource.from_file(file_path).delete()
-    success(f"Deleted: {title}")
-
-
-# =============================================================================
-# Form commands
-# =============================================================================
-
-
-@docs.section("resource")
-@docs.maturity("unstable")
-@click.group()
-def form():
-    """Google Forms operations"""
-    pass
-
-
-@form.command("clone")
-@click.argument("url")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <title>.form.gax.md)",
-)
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    type=click.Choice(["md", "yaml"]),
-    default="md",
-    help="Content format: md (readable, default) or yaml (round-trip safe)",
-)
-@handle_errors
-def form_clone(url, output, fmt):
-    """Clone a Google Form to a local .form.gax.md file.
-
-    By default, creates a human-readable markdown representation.
-    Use --format yaml for faithful round-trip representation (required for push).
-    """
-    from .ui import success
-
-    file_path = Form.from_url(url).clone(output=output, format=fmt)
-    success(f"Created: {file_path}")
-    if fmt == "md":
-        click.echo("Note: Use --format yaml for round-trip safe format")
-
-
-@form.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def form_pull(file):
-    """Pull latest form definition from Google Forms."""
-    from .ui import success
-
-    Form.from_file(file).pull()
-    success(f"Updated: {file}")
-
-
-@form.command("plan")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    default="form.plan.yaml",
-    help="Output plan file",
-)
-@handle_errors
-def form_plan(file, output):
-    """Preview form changes (diff)."""
-    diff_text = Form.from_file(file).diff()
-    if diff_text is None:
-        click.echo("No changes to apply.")
-        return
-    click.echo(diff_text)
-
-
-@form.command("apply")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
-@handle_errors
-def form_apply(file, yes):
-    """Apply form changes to Google Forms."""
-    _confirm_and_push(Form.from_file(file), yes=yes)
-
-
-# =============================================================================
-# Doc commands
-# =============================================================================
-
-
-@docs.section("resource")
-@click.group()
-def doc():
-    """Google Docs operations"""
-    pass
-
-
-@doc.group("tab")
-def doc_tab():
-    """Single tab operations"""
-    pass
-
-
-@doc_tab.command("list")
-@click.argument("url")
-@handle_errors
-def doc_tab_list(url: str):
-    """List tabs in a document (TSV output)."""
-    Doc.from_url(url).tab_list(sys.stdout)
-
-
-@doc_tab.command("import")
-@click.argument("url")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output tracking file (default: <filename>.tab.gax.md)",
-)
-@handle_errors
-def doc_tab_import(url: str, file: Path, output: Path | None):
-    """Import a markdown file as a new tab in a document."""
-    from .ui import success
-
-    tracking_path = Doc.from_url(url).tab_import(file, output=output)
-    success(f"Created: {tracking_path}")
-
-
-@doc_tab.command("clone")
-@click.argument("url")
-@click.argument("tab_name")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <tab>.tab.gax.md)",
-)
-@handle_errors
-def doc_tab_clone(url: str, tab_name: str, output: Path | None):
-    """Clone a single tab to a .tab.gax.md file."""
-    from .ui import success
-
-    file_path = Tab.from_url(url).clone(output=output, tab_name=tab_name)
-    success(f"Created: {file_path}")
-
-
-@doc_tab.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def doc_tab_pull(file: Path):
-    """Pull latest content for a single tab."""
-    from .ui import success
-
-    Tab.from_file(file).pull()
-    success(f"Updated: {file}")
-
-
-@doc_tab.command("diff")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def doc_tab_diff(file: Path):
-    """Show diff between local file and remote tab."""
-    diff_text = Tab.from_file(file).diff()
-    if diff_text is None:
-        click.echo("No differences.")
-    else:
-        click.echo(diff_text)
-
-
-@doc_tab.command("push")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation prompt")
-@click.option(
-    "--patch",
-    "use_patch",
-    is_flag=True,
-    help="Incremental push: apply only changed elements (experimental)",
-)
-@handle_errors
-def doc_tab_push(file: Path, yes: bool, use_patch: bool):
-    """Push local changes to a single tab (with confirmation).
-
-    The default push path is full-replace (see ADR 023). The ``--patch`` flag
-    selects an **experimental** incremental push path (ADR 027) that diffs the
-    local markdown against the live document and applies only the changed
-    elements. The ``--patch`` path is under evaluation and may fail on
-    structural changes; when in doubt, omit the flag.
-    """
-    from .gdoc import parse_multipart, extract_doc_id
-    from .ui import success, error
-
-    t = Tab.from_file(file)
-
-    if use_patch:
-        from .gdoc.diff_push import preview_diff
-        from .gdoc import native_md as _native_md
-
-        section = parse_multipart(file.read_text(encoding="utf-8"))[0]
-        source_url = section.source
-        tab_name = section.section_title
-        document_id = extract_doc_id(source_url)
-
-        content_to_push = _native_md.inline_images_from_store(section.content)
-
-        preview = preview_diff(document_id, tab_name, content_to_push)
-
-        if not preview.ops:
-            click.echo("No differences to push.")
-            return
-
-        click.echo("Patch operations:")
-        click.echo("-" * 40)
-        for line in preview.summary_lines:
-            click.echo(line)
-        click.echo("-" * 40)
-
-        if preview.warnings:
-            for w in preview.warnings:
-                error(w)
-            click.echo("Use regular push (without --patch) for structural changes.")
-            sys.exit(1)
-
-        if not yes:
-            if not click.confirm("Apply patch?"):
-                click.echo("Aborted.")
-                return
-
-        t.push(use_patch=True)
-        success("Patched successfully.")
-    else:
-        diff_text = t.diff()
-        if diff_text is None:
-            click.echo("No differences to push.")
-            return
-
-        click.echo("Changes to push:")
-        click.echo("-" * 40)
-        click.echo(diff_text)
-        click.echo("-" * 40)
-
-        from .gdoc.ir import from_markdown, check_unsupported
-
-        section = parse_multipart(file.read_text(encoding="utf-8"))[0]
-        push_warnings = check_unsupported(from_markdown(section.content))
-        for w in push_warnings:
-            click.echo(f"  Warning: {w.feature}: {w.detail}")
-
-        click.echo(
-            "Warning: markdown cannot faithfully represent a Google Doc. "
-            "Non-markdown formatting (colors, fonts, alignment, comments, "
-            "suggestions, images) may be lost. Use --patch for incremental "
-            "updates that preserve formatting (experimental)."
-        )
-
-        if not yes:
-            if not click.confirm("Push these changes?"):
-                click.echo("Aborted.")
-                return
-
-        t.push()
-        success("Pushed successfully.")
-
-
-@doc.command("clone")
-@click.argument("url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output file (default: <title>.doc.gax.md)",
-)
-@click.option(
-    "--with-comments",
-    is_flag=True,
-    help="Include document comments as separate sections",
-)
-@click.option(
-    "-q",
-    "--quiet",
-    is_flag=True,
-    help="Suppress multi-tab status message",
-)
-@handle_errors
-def doc_clone(url: str, output: Path | None, with_comments: bool, quiet: bool):
-    """Clone a Google Doc to a local .doc.gax.md file.
-
-    Clones a single tab. For multi-tab documents, use 'gax doc checkout'.
-    """
-    from .gdoc import extract_doc_id, get_tabs_list
-    from .ui import success
-
-    file_path = Tab.from_url(url).clone(output=output, with_comments=with_comments)
-    success(f"Created: {file_path}")
-
-    if not quiet:
-        document_id = extract_doc_id(url)
-        tabs = get_tabs_list(document_id)
-        if len(tabs["tabs"]) > 1:
-            first_tab = tabs["tabs"][0].title
-            click.echo(
-                f'  Tab "{first_tab}" cloned (1 of {len(tabs["tabs"])} tabs).\n'
-                f"  For all tabs: gax doc checkout {url}"
-            )
-
-
-@doc.command("pull")
-@click.argument("file", type=click.Path(exists=True, path_type=Path))
-@click.option(
-    "--with-comments",
-    is_flag=True,
-    help="Include document comments as separate sections",
-)
-@handle_errors
-def doc_pull(file: Path, with_comments: bool):
-    """Pull latest content from Google Docs to local file."""
-    from .ui import success
-
-    Tab.from_file(file).pull(with_comments=with_comments)
-    success(f"Updated: {file}")
-
-
-@doc.command("checkout")
-@click.argument("url")
-@click.option(
-    "-o",
-    "--output",
-    type=click.Path(path_type=Path),
-    help="Output folder (default: <title>.doc.gax.md.d)",
-)
-@handle_errors
-def doc_checkout(url: str, output: Path | None):
-    """Checkout all tabs to individual files in a folder.
-
-    Creates a folder with individual .doc.gax.md files for each tab.
-    """
-    from .ui import success
-
-    folder = Doc.from_url(url).checkout(output=output)
-    success(f"Checked out to: {folder}")
-
-
-# =============================================================================
-# Slides commands
-# =============================================================================
-
-
-@docs.section("resource")
-@main.group()
-def slides():
-    """Google Slides operations"""
-    pass
-
-
-@slides.command("checkout")
-@click.argument("url")
-@click.option(
-    "--output",
-    "-o",
-    type=click.Path(path_type=Path),
-    help="Output directory path",
-)
-@click.option(
-    "-f",
-    "--format",
-    "fmt",
-    default="md",
-    type=click.Choice(["md", "json"]),
-    show_default=True,
-    help="Output format: md (read-only) or json (read-write)",
-)
-@handle_errors
-def slides_checkout(url: str, output: Path | None, fmt: str):
-    """Checkout a Google Slides presentation to a local directory.
-
-    Creates a .slides.gax.md.d/ folder with one file per slide.
-
-    \b
-    Formats:
-        md   — human-readable markdown (pull only, no push)
-        json — full-fidelity JSON (supports push)
-
-    \b
-    Examples:
-        gax slides checkout https://docs.google.com/presentation/d/abc123/edit
-        gax slides checkout abc123 -o my_slides
-        gax slides checkout abc123 --format json
-    """
-    from .ui import success
-
-    folder_path = Presentation.from_url(url).checkout(output=output, fmt=fmt)
-    success(f"Checked out: {folder_path}")
-
-
-@slides.command("pull")
-@click.argument("path", type=click.Path(exists=True, path_type=Path))
-@handle_errors
-def slides_pull(path: Path):
-    """Pull latest slides from Google.
-
-    Works with both single .slides.gax.md files and .slides.gax.md.d/ folders.
-
-    \b
-    Examples:
-        gax slides pull my_deck.slides.gax.md.d/
-        gax slides pull 00_Welcome.slides.gax.md
-    """
-    from .ui import success
-
-    if path.is_dir():
-        Presentation.from_file(path).pull()
-    else:
-        Slide.from_file(path).pull()
-    success(f"Updated: {path}")
-
-
-@slides.command("push")
-@click.argument("path", type=click.Path(exists=True, path_type=Path))
-@click.option("-y", "--yes", is_flag=True, help="Skip confirmation prompt")
-@handle_errors
-def slides_push(path: Path, yes: bool):
-    """Push local slides to Google. JSON format only.
-
-    Markdown checkouts cannot be pushed — re-checkout with --format json.
-    Works with both single .slides.gax.md files and .slides.gax.md.d/ folders.
-
-    \b
-    Examples:
-        gax slides push my_deck.slides.gax.md.d/
-        gax slides push my_deck.slides.gax.md.d/ -y
-        gax slides push 00_Welcome.slides.gax.md
-    """
-    from .ui import success
-
-    if path.is_dir():
-        p = Presentation.from_file(path)
-        diff_text = p.diff()
-        if diff_text is None:
-            click.echo("No changes to push.")
-            return
-        if not yes:
-            click.echo(diff_text)
-            if not click.confirm("Push these changes?"):
-                click.echo("Cancelled.")
-                return
-        p.push()
-    else:
-        s = Slide.from_file(path)
-        diff_text = s.diff()
-        if diff_text is None:
-            click.echo("No changes to push.")
-            return
-        if not yes:
-            click.echo(diff_text)
-            if not click.confirm("Push these changes?"):
-                click.echo("Cancelled.")
-                return
-        s.push()
-
-    success(f"Pushed: {path}")
-
-
-# Register command groups
+main.add_command(sheet)
 main.add_command(doc)
 main.add_command(mail_group, name="mail")
 main.add_command(mailbox_group, name="mailbox")
@@ -2388,7 +384,12 @@ main.add_command(form)
 main.add_command(draft)  # Flattened from mail.draft (ADR 020)
 main.add_command(contacts)
 main.add_command(file_group, name="file")
+main.add_command(slides)
 
+
+# =============================================================================
+# Utility commands
+# =============================================================================
 
 REPO = "HeinrichHartmann/gax"
 ISSUES_URL = f"https://github.com/{REPO}/issues"
