+++ gax/draft/__init__.py
@@ -1,0 +1,16 @@
+"""Gmail draft management for gax.
+
+Re-exports from draft.py.
+"""
+
+from .draft import (  # noqa: F401
+    DraftHeader,
+    parse_draft,
+    format_draft,
+    parse_draft_id,
+    get_header,
+    build_message,
+    extract_body,
+    fetch_draft,
+    Draft,
+)
