+++ gax/draft/draft.py
@@ -65,9 +65,9 @@ from typing import Any
 
 from googleapiclient.discovery import build
 
-from .auth import get_authenticated_credentials
-from . import multipart
-from .resource import Resource
+from ..auth import get_authenticated_credentials
+from .. import multipart
+from ..resource import Resource
 
 logger = logging.getLogger(__name__)
