+++ gax/filter/__init__.py
@@ -1,0 +1,24 @@
+"""Gmail filter management for gax.
+
+Re-exports from filter.py.
+"""
+
+from .filter import (  # noqa: F401
+    FilterHeader,
+    parse_filters_file,
+    format_filters_file,
+    CRITERIA_KEYS,
+    get_service,
+    fetch_filters,
+    fetch_label_maps,
+    api_to_criteria,
+    criteria_to_api,
+    api_to_action,
+    action_to_api,
+    get_or_create_label,
+    criteria_hash,
+    generate_filter_name,
+    compute_changes,
+    format_diff_summary,
+    Filter,
+)
