+++ gax/gcal/gcal.py
@@ -45,8 +45,8 @@ from typing import Any, Optional
 import yaml
 from googleapiclient.discovery import build
 
-from .auth import get_authenticated_credentials
-from .resource import Resource
+from ..auth import get_authenticated_credentials
+from ..resource import Resource
 
 logger = logging.getLogger(__name__)
