+++ gax/gdrive/__init__.py
@@ -1,0 +1,21 @@
+"""Google Drive file and folder operations for gax.
+
+Re-exports from gdrive.py.
+"""
+
+from .gdrive import (  # noqa: F401
+    create_tracking_file,
+    read_tracking_file,
+    extract_file_id,
+    download_file,
+    upload_file,
+    update_file,
+    set_public,
+    WORKSPACE_MIME_TYPES,
+    FOLDER_MIME_TYPE,
+    extract_folder_id,
+    get_folder_metadata,
+    list_folder,
+    File,
+    Folder,
+)
