+++ gax/gslides/gslides.py
@@ -32,9 +32,9 @@ from pathlib import Path
 import yaml
 from googleapiclient.discovery import build
 
-from .auth import get_authenticated_credentials
-from .multipart import Section, format_section, parse_multipart
-from .resource import Resource
+from ..auth import get_authenticated_credentials
+from ..multipart import Section, format_section, parse_multipart
+from ..resource import Resource
 
 logger = logging.getLogger(__name__)
