+++ gax/gtask/__init__.py
@@ -1,0 +1,25 @@
+"""Google Tasks sync for gax.
+
+Re-exports from gtask.py.
+"""
+
+from .gtask import (  # noqa: F401
+    TaskItem,
+    get_tasks_service,
+    list_tasklists,
+    list_tasks,
+    get_task,
+    create_task,
+    update_task,
+    delete_task,
+    api_to_task,
+    task_to_api_body,
+    task_to_yaml,
+    yaml_to_task,
+    format_tasks_md,
+    parse_tasks_md,
+    format_tasks_yaml,
+    resolve_tasklist_id,
+    TaskList,
+    Task,
+)
