+++ tests/test_gtask.py
@@ -353,8 +353,8 @@ class TestTaskDiff:
         assert "New task" in result
         assert "2026-04-21" in result
 
-    @patch("gax.gtask.get_task")
-    @patch("gax.gtask.api_to_task")
+    @patch("gax.gtask.gtask.get_task")
+    @patch("gax.gtask.gtask.api_to_task")
     def test_no_changes_returns_none(self, mock_api_to, mock_get, tmp_path):
         """Identical local/remote returns None."""
         task = TaskItem("A", "TL1", "", "", "Same", "needsAction", due="2026-04-21")
@@ -369,8 +369,8 @@ class TestTaskDiff:
         result = Task(path=path).diff()
         assert result is None
 
-    @patch("gax.gtask.get_task")
-    @patch("gax.gtask.api_to_task")
+    @patch("gax.gtask.gtask.get_task")
+    @patch("gax.gtask.gtask.api_to_task")
     def test_field_changes(self, mock_api_to, mock_get, tmp_path):
         """Changed fields shown in diff output."""
         local = TaskItem(
@@ -398,7 +398,7 @@ class TestTaskDiff:
 class TestTaskDone:
     """Tests for Task.done() shortcut."""
 
-    @patch("gax.gtask.update_task")
+    @patch("gax.gtask.gtask.update_task")
     def test_done_sets_completed_and_pushes(self, mock_update, tmp_path):
         """done() sets status=completed, writes file, calls API."""
         task = TaskItem("A", "TL1", "", "", "My task", "needsAction")
